Moved the "data" and "implementation" files into separate subdirectories.
The aim here is to (hopefully) make it easier to navigate the directory
structure of the header files. So that the most important files could
be found as easily as possible.
