+++ core/CMakeLists.txt
@@ -8,30 +8,29 @@
 vecmem_add_library( vecmem_core core SHARED
    # STL mimicking containers.
    "include/vecmem/containers/array.hpp"
-   "include/vecmem/containers/array.ipp"
+   "include/vecmem/containers/impl/array.ipp"
    "include/vecmem/containers/const_device_array.hpp"
    "include/vecmem/containers/const_device_vector.hpp"
    "include/vecmem/containers/device_array.hpp"
-   "include/vecmem/containers/device_array.ipp"
+   "include/vecmem/containers/impl/device_array.ipp"
    "include/vecmem/containers/device_vector.hpp"
-   "include/vecmem/containers/device_vector.ipp"
+   "include/vecmem/containers/impl/device_vector.ipp"
    "include/vecmem/containers/static_vector.hpp"
-   "include/vecmem/containers/static_vector.ipp"
-   "include/vecmem/containers/vector.hpp"
-   "include/vecmem/containers/vector.ipp"
-   # Jagged vectors.
-   "include/vecmem/containers/jagged_vector.hpp"
-   "include/vecmem/containers/jagged_vector_data.hpp"
-   "include/vecmem/containers/jagged_vector_data.ipp"
+   "include/vecmem/containers/impl/static_vector.ipp"
    "include/vecmem/containers/jagged_device_vector.hpp"
-   "include/vecmem/containers/jagged_device_vector.ipp"
-   "include/vecmem/containers/details/jagged_vector_view.hpp"
-   "include/vecmem/containers/details/jagged_vector_view.ipp"
-   # Implementation details for the containers.
-   "include/vecmem/containers/details/vector_buffer.hpp"
-   "include/vecmem/containers/details/vector_buffer.ipp"
-   "include/vecmem/containers/details/vector_view.hpp"
-   "include/vecmem/containers/details/vector_view.ipp"
+   "include/vecmem/containers/impl/jagged_device_vector.ipp"
+   "include/vecmem/containers/jagged_vector.hpp"
+   "include/vecmem/containers/vector.hpp"
+   "include/vecmem/containers/impl/vector.ipp"
+   # Data holding/transporting types.
+   "include/vecmem/containers/data/jagged_vector_data.hpp"
+   "include/vecmem/containers/impl/jagged_vector_data.ipp"
+   "include/vecmem/containers/data/jagged_vector_view.hpp"
+   "include/vecmem/containers/impl/jagged_vector_view.ipp"
+   "include/vecmem/containers/data/vector_buffer.hpp"
+   "include/vecmem/containers/impl/vector_buffer.ipp"
+   "include/vecmem/containers/data/vector_view.hpp"
+   "include/vecmem/containers/impl/vector_view.ipp"
    # Allocator
    "include/vecmem/memory/allocator.hpp"
    "include/vecmem/memory/allocator.ipp"
