+++ core/include/vecmem/containers/impl/array.ipp
