+++ core/include/vecmem/containers/impl/device_array.ipp
