+++ core/include/vecmem/containers/impl/device_vector.ipp
