+++ core/include/vecmem/containers/impl/static_vector.ipp
