+++ core/include/vecmem/containers/impl/vector_buffer.ipp
