+++ core/include/vecmem/containers/impl/vector_view.ipp
