+++ tests/cuda/test_cuda_jagged_vector_view.cpp
@@ -11,7 +11,7 @@
 #include "vecmem/memory/cuda/managed_memory_resource.hpp"
 #include "vecmem/containers/vector.hpp"
 #include "vecmem/containers/jagged_vector.hpp"
-#include "vecmem/containers/jagged_vector_data.hpp"
+#include "vecmem/containers/data/jagged_vector_data.hpp"
 #include "vecmem/containers/jagged_device_vector.hpp"
 
 #include <gtest/gtest.h>
