+++ tests/cuda/test_cuda_jagged_vector_view_kernels.cuh
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include "vecmem/containers/details/jagged_vector_view.hpp"
+#include "vecmem/containers/data/jagged_vector_view.hpp"
 
 void doubleJagged(
     vecmem::details::jagged_vector_view<int> & jag
